backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
redis==5.2.0
openai==1.54.3
tiktoken==0.8.0
supabase==2.9.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.36.0
google-api-python-client==2.152.0
razorpay==1.4.2
python-multipart==0.0.17
